--- com.unity.netcode.gameobjects/Tests/Editor/NetworkManagerMessageHandlerTests.cs
@@ -1,52 +1,0 @@
-using System;
-
-
-namespace Unity.Netcode.EditorTests
-{
-    internal class DummyTransport : TestingNetworkTransport
-    {
-        public override ulong ServerClientId { get; } = 0;
-        public override void Send(ulong clientId, ArraySegment<byte> payload, NetworkDelivery networkDelivery)
-        {
-        }
-
-        public override NetworkEvent PollEvent(out ulong clientId, out ArraySegment<byte> payload, out float receiveTime)
-        {
-            clientId = 0;
-            payload = new ArraySegment<byte>();
-            receiveTime = 0;
-            return NetworkEvent.Nothing;
-        }
-
-        public override bool StartClient()
-        {
-            return true;
-        }
-
-        public override bool StartServer()
-        {
-            return true;
-        }
-
-        public override void DisconnectRemoteClient(ulong clientId)
-        {
-        }
-
-        public override void DisconnectLocalClient()
-        {
-        }
-
-        public override ulong GetCurrentRtt(ulong clientId)
-        {
-            return 0;
-        }
-
-        public override void Shutdown()
-        {
-        }
-
-        public override void Initialize(NetworkManager networkManager = null)
-        {
-        }
-    }
-}
