--- com.unity.netcode.gameobjects/Tests/Editor/NetworkManagerMessageHandlerTests.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 976ca592c7fa4bcb854203dfbadc0ad9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
