+++ com.unity.netcode.gameobjects/Tests/Runtime/NetworkManagerCustomMessageManagerTests.cs
@@ -1,28 +1,34 @@
 using NUnit.Framework;
 using UnityEngine;
 
-namespace Unity.Netcode.EditorTests
+namespace Unity.Netcode.RuntimeTests
 {
-    public class NetworkManagerSceneManagerTests
+    public class NetworkManagerCustomMessageManagerTests
     {
         [Test]
-        public void SceneManagerAssigned()
+        public void CustomMessageManagerAssigned()
         {
-            var gameObject = new GameObject(nameof(SceneManagerAssigned));
+            var gameObject = new GameObject(nameof(CustomMessageManagerAssigned));
             var networkManager = gameObject.AddComponent<NetworkManager>();
             var transport = gameObject.AddComponent<DummyTransport>();
+
             networkManager.NetworkConfig = new NetworkConfig();
+
+
+
             // Set dummy transport that does nothing
             networkManager.NetworkConfig.NetworkTransport = transport;
 
-            NetworkSceneManager preManager = networkManager.SceneManager;
+            CustomMessagingManager preManager = networkManager.CustomMessagingManager;
 
-            // Start server to cause initialization process
+            // Start server to cause initialization
             networkManager.StartServer();
 
             Debug.Assert(preManager == null);
-            Debug.Assert(networkManager.SceneManager != null);
+            Debug.Assert(networkManager.CustomMessagingManager != null);
 
+
+            networkManager.Shutdown();
             Object.DestroyImmediate(gameObject);
         }
     }
