+++ com.unity.netcode.gameobjects/Tests/Runtime/NetworkManagerSceneManagerTests.cs
@@ -1,28 +1,27 @@
 using NUnit.Framework;
 using UnityEngine;
 
-namespace Unity.Netcode.EditorTests
+namespace Unity.Netcode.RuntimeTests
 {
-    public class NetworkManagerCustomMessageManagerTests
+    public class NetworkManagerSceneManagerTests
     {
         [Test]
-        public void CustomMessageManagerAssigned()
+        public void SceneManagerAssigned()
         {
-            var gameObject = new GameObject(nameof(CustomMessageManagerAssigned));
+            var gameObject = new GameObject(nameof(SceneManagerAssigned));
             var networkManager = gameObject.AddComponent<NetworkManager>();
             var transport = gameObject.AddComponent<DummyTransport>();
-
             networkManager.NetworkConfig = new NetworkConfig();
             // Set dummy transport that does nothing
             networkManager.NetworkConfig.NetworkTransport = transport;
 
-            CustomMessagingManager preManager = networkManager.CustomMessagingManager;
+            NetworkSceneManager preManager = networkManager.SceneManager;
 
-            // Start server to cause initialization
+            // Start server to cause initialization process
             networkManager.StartServer();
 
             Debug.Assert(preManager == null);
-            Debug.Assert(networkManager.CustomMessagingManager != null);
+            Debug.Assert(networkManager.SceneManager != null);
 
             Object.DestroyImmediate(gameObject);
         }
