+++ com.unity.netcode.gameobjects/Tests/Runtime/StartStopTests.cs
@@ -1,7 +1,7 @@
 using NUnit.Framework;
 using UnityEngine;
 
-namespace Unity.Netcode.EditorTests
+namespace Unity.Netcode.RuntimeTests
 {
     public class StartStopTests
     {
